1. Доработаны расщирения для CacheRepository.

+++ packages/front/src/implementation/repository/CacheRepository.ts
@@ -1,38 +1,47 @@
-import { IEntity } from "@borodindmitriy/interfaces";
-import { getErrorMessage } from "@borodindmitriy/utils";
+import { IForm } from "@borodindmitriy/interfaces";
+import { EventEmitter } from "@borodindmitriy/isomorphic";
+import { getErrorMessage, isString } from "@borodindmitriy/utils";
 import { action, computed, observable, ObservableMap, runInAction } from "mobx";
 import { ICacheRepository } from "../../interfaces/repository/ICacheRepository";
 
-export class CacheRepository<E extends IEntity> implements ICacheRepository<E> {
+// tslint:disable: object-literal-sort-keys
+
+export class CacheRepository<T extends IForm> extends EventEmitter implements ICacheRepository<T> {
+  public static events = {
+    update: `[ Repository ][ UPDATE ]`,
+    remove: `[ Repository ][ REMOVE ]`,
+  };
+
   @observable
-  protected collection: ObservableMap<string, E>;
+  protected collection: ObservableMap<string, T>;
+  protected Entity: new (...args: any[]) => T;
 
-  protected Entity: new (data?: any) => E;
+  constructor(Entity: new (...args: any[]) => T) {
+    super();
 
-  constructor(Entity: new (data?: any) => E) {
     // * DEPS
     this.Entity = Entity;
 
     // ! OBSERVABLE
-    runInAction(`[ ${this.constructor.name} ][ SET_INITIAL_VALUE ]`, () => {
-      this.collection = observable.map();
-    });
+    runInAction(`[ ${this.constructor.name} ][ SET_INITIAL_VALUE ]`, () => (this.collection = observable.map()));
 
     // * BINDINGS
     this.update = this.update.bind(this);
     this.remove = this.remove.bind(this);
     this.destroy = this.destroy.bind(this);
     this.filter = this.filter.bind(this);
+    this.collectionDidUpdate = this.collectionDidUpdate.bind(this);
+    this.collectionDidRemove = this.collectionDidRemove.bind(this);
   }
 
   @computed({ name: "[ CACHE ][ REPOSITORY ][ MAP ]" })
-  get map(): ObservableMap<string, E> {
+  get map(): ObservableMap<string, T> {
     return this.collection;
   }
 
   @computed({ name: "[ CACHE ][ REPOSITORY ][ LIST ]" })
-  get list(): E[] {
-    const list: E[] = [];
+  get list(): T[] {
+    const list: T[] = [];
 
     for (const value of this.collection.values()) {
       list.push(value);
@@ -42,11 +51,19 @@ export class CacheRepository<E extends IEntity> implements ICacheRepository<E> {
   }
 
   @action("[ CACHE ][ REPOSITORY ][ UPDATE ]")
-  public update(entities: E[]): void {
+  public update(items: T[]): void {
     try {
-      for (const entity of entities) {
-        this.collection.set(entity.id, entity);
+      const entities: T[] = [];
+
+      for (const item of items) {
+        if (item instanceof this.Entity && isString(item.id)) {
+          this.collection.set(item.id, item);
+          entities.push(item);
+        }
       }
+
+      this.collectionDidUpdate(entities);
+      this.emit(CacheRepository.events.update, entities);
     } catch (error) {
       console.error(`[ ${this.constructor.name} ][ UPDATE ][ ${getErrorMessage(error)} ]`);
     }
@@ -55,9 +72,19 @@ export class CacheRepository<E extends IEntity> implements ICacheRepository<E> {
   @action("[ CACHE ][ REPOSITORY ][ REMOVE ]")
   public remove(ids: string[]): void {
     try {
+      const entities: T[] = [];
+
       for (const id of ids) {
-        this.collection.delete(id);
+        const entity: T | void = this.collection.get(id);
+
+        if (entity instanceof this.Entity) {
+          entities.push(entity);
+          this.collection.delete(id);
+        }
       }
+
+      this.collectionDidRemove(entities);
+      this.emit(CacheRepository.events.update, entities);
     } catch (error) {
       console.error(`[ ${this.constructor.name} ][ REMOVE ][ ${getErrorMessage(error)} ]`);
     }
@@ -66,9 +93,22 @@ export class CacheRepository<E extends IEntity> implements ICacheRepository<E> {
   @action("[ CACHE ][ REPOSITORY ][ DESTROY ]")
   protected destroy(): void {
     this.collection.clear();
+    this.collectionDidDestroied();
   }
 
-  protected filter(list: E[]): E[] {
+  protected filter(list: T[]): T[] {
     return list;
+  }
+
+  protected collectionDidUpdate(entities: T[]): void {
+    // ! HOOK FOR COLLECTION UPDATE
+  }
+
+  protected collectionDidRemove(entities: T[]): void {
+    // ! HOOK FOR COLLECTION UPDATE
+  }
+
+  protected collectionDidDestroied(): void {
+    // ! HOOK FOR COLLECTION DESTROY
   }
 }
